Added groceries, reworked the error handling

+++ src/logic/operations/FlagOperations.go
@@ -1,15 +1,21 @@
 package operations
 
 import (
+	"mealwhile/errors"
 	"mealwhile/logic/model"
 )
 
 type FlagOperations struct {
-	Service CrudService
+	Service           CrudService
+	groceryOperations *GroceryOperations
 }
 
 func NewFlagOperations(service CrudService) FlagOperations {
 	return FlagOperations{Service: service}
+}
+
+func (o *FlagOperations) SetGroceryOperations(groceryOps *GroceryOperations) {
+	o.groceryOperations = groceryOps
 }
 
 func (ops FlagOperations) Create(entity model.CrudEntity) (model.CrudEntity, error) {
@@ -29,5 +35,16 @@ func (ops FlagOperations) Update(entity model.CrudEntity) (model.CrudEntity, err
 }
 
 func (ops FlagOperations) Delete(target model.CrudEntity, id string) error {
+	// Check if the flag to be deleted is contained in a grocery
+	referenced, err := ops.groceryOperations.FlagReferenced(id)
+
+	if err != nil {
+		return err
+	}
+
+	if referenced {
+		return errors.NewBadRequest("The flag cannot be deleted, as it is used by a grocery")
+	}
+
 	return ops.Service.Delete(target, id)
 }
